deepscan-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy[asyncio]==2.0.37
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-nmap==0.7.1
scapy==2.6.1
mac-vendor-lookup==0.1.12
python-dateutil==2.9.0.post0

